requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.6.2.post1
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==24.2.0
Automat==24.8.1
beautifulsoup4==4.12.3
blinker==1.8.2
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
comm==0.2.2
comtypes==1.4.7
confection==0.1.5
constantly==23.10.4
cryptography==43.0.3
cssselect==1.2.0
cymem==2.0.8
dataclasses-json==0.6.7
db==0.1.1
db-sqlite3==0.0.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
Django==5.1.2
django-environ==0.11.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
executing==2.2.0
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.29.5
ipython==8.32.0
itemadapter==0.9.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.7.0
langcodes==3.4.1
language_data==1.2.0
libclang==18.1.1
llama-cloud==0.1.16
llama-cloud-services==0.6.8
llama-index==0.12.25
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.1
llama-index-core==0.12.25
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.9
llama-index-llms-openai==0.3.27
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.7
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.4.post1
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
MouseInfo==0.1.3
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==2.0.2
ollama==0.4.7
openai==1.68.2
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parsel==1.9.1
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
playsound3==2.5.2
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.2.0
Protego==0.3.1
protobuf==5.29.0
psutil==6.1.1
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyDispatcher==2.0.7
PyGetWindow==0.0.9
Pygments==2.18.0
pyjokes==0.8.3
PyMsgBox==1.0.9
pyOpenSSL==24.2.1
pypdf==5.4.0
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
queuelib==1.7.0
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
rich==13.9.2
safetensors==0.4.5
secure-smtplib==0.1.1
selenium==4.26.1
service-identity==24.2.0
setuptools==75.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.11.0
SQLAlchemy==2.0.39
sqlparse==0.5.1
srsly==2.4.8
stack-data==0.6.3
striprtf==0.0.26
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
thinc==8.3.2
tiktoken==0.9.0
tldextract==5.1.3
tokenizers==0.20.1
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.0
trio==0.27.0
trio-websocket==0.11.1
Twisted==24.10.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
w3lib==2.2.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.1
wheel==0.45.1
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.16.0
zope.interface==7.1.1
